requirements.txt
flask==0.10.1
werkzeug==0.9.6
flask-sqlalchemy==2.0
psycopg2==2.5.4
lxml==3.3.6
requests==2.4.0